+++ packages/reown_appkit/lib/modal/pages/preview_send/preview_send_solana.dart
@@ -120,17 +120,8 @@ class _PreviewSendSolanaState extends State<PreviewSendSolana> {
 
   Future<solana.Transaction> _contructSolanaTX(double valueToSend) async {
     // Create a connection to the devnet cluster.
-    final chainId = ReownAppKitModalNetworks.getIdFromChain(
-      _sendTokenData.chainId!,
-    );
-    final chainInfo = ReownAppKitModalNetworks.getNetworkInfo(
-      'solana',
-      chainId,
-    );
-    // Create a connection to the devnet cluster.
     final appKitModal = ModalProvider.of(context).instance;
-    final projectId = appKitModal.appKit!.core.projectId;
-    final rpcUrl = chainInfo!.formattedRpcUrl(projectId);
+    final rpcUrl = appKitModal.selectedChain!.extraRpcUrls.first;
     final cluster = solana.Cluster.https(Uri.parse(rpcUrl).authority);
     // final cluster = solana.Cluster.devnet;
     final connection = solana.Connection(cluster);
@@ -183,18 +174,9 @@ class _PreviewSendSolanaState extends State<PreviewSendSolana> {
 
   Future<void> _estimateNetworkCost() async {
     try {
-      final chainId = ReownAppKitModalNetworks.getIdFromChain(
-        _sendTokenData.chainId!,
-      );
-      final chainInfo = ReownAppKitModalNetworks.getNetworkInfo(
-        'solana',
-        chainId,
-      );
-
       // Create a connection to the devnet cluster.
       final appKitModal = ModalProvider.of(context).instance;
-      final projectId = appKitModal.appKit!.core.projectId;
-      final rpcUrl = chainInfo!.formattedRpcUrl(projectId);
+      final rpcUrl = appKitModal.selectedChain!.extraRpcUrls.first;
       final cluster = solana.Cluster.https(Uri.parse(rpcUrl).authority);
       // final cluster = solana.Cluster.devnet;
       final connection = solana.Connection(cluster);
@@ -268,6 +250,10 @@ class _PreviewSendSolanaState extends State<PreviewSendSolana> {
           },
         ),
       );
+      if (response is Map && response.containsKey('errorCode')) {
+        throw Exception('${response['errorMessage']}');
+      }
+
       _analyticsService.sendEvent(
         WalletFeatureSendSuccess(
           network: _sendTokenData.chainId!,
@@ -292,8 +278,20 @@ class _PreviewSendSolanaState extends State<PreviewSendSolana> {
       );
     } on Exception catch (e) {
       _toastService.show(
-        ToastMessage(type: ToastType.error, text: e.toString()),
+        ToastMessage(
+          type: ToastType.error,
+          text: e.toString().replaceFirst('Exception: ', ''),
+        ),
       );
+      _analyticsService.sendEvent(
+        WalletFeatureSendError(
+          network: _sendTokenData.chainId!,
+          sendToken: _sendTokenData.symbol!,
+          sendAmount: valueToSend,
+        ),
+      );
+    } on ReownSignError catch (e) {
+      _toastService.show(ToastMessage(type: ToastType.error, text: e.message));
       _analyticsService.sendEvent(
         WalletFeatureSendError(
           network: _sendTokenData.chainId!,